+++ package.json
@@ -16,34 +16,36 @@
   ],
   "dependencies": {
     "blacklist": "^1.1.4",
-    "prop-types": "^15.5.10"
+    "prop-types": "^15.7.2"
   },
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.25.0",
-    "babel-plugin-transform-react-jsx": "^6.24.1",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-react": "^6.24.1",
-    "coveralls": "^2.13.1",
-    "eslint-config-xo-react": "^0.13.0",
-    "eslint-plugin-react": "^7.1.0",
-    "ink": "^0.2.1",
-    "jest": "^20.0.4",
-    "xo": "^0.18.2"
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/node": "^7.2.2",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/preset-react": "^7.0.0",
+    "ava": "^1.3.1",
+    "babel-eslint": "^10.0.1",
+    "coveralls": "^3.0.3",
+    "eslint-config-xo-react": "^0.19.0",
+    "eslint-plugin-react": "^7.12.4",
+    "eslint-plugin-react-hooks": "^1.6.0",
+    "ink": "^2.1.1",
+    "jest": "^24.7.1",
+    "react": "^16.8.6",
+    "xo": "^0.24.0"
   },
   "babel": {
-    "presets": [
-      "babel-preset-react"
-    ],
     "plugins": [
-      [
-        "transform-react-jsx",
-        {
-          "pragma": "h",
-          "useBuiltIns": true
-        }
-      ]
+      "@babel/plugin-proposal-class-properties"
+    ],
+    "presets": [
+      "@ava/stage-4",
+      "@babel/preset-react"
     ]
+  },
+  "ava": {
+    "color": true
   },
   "xo": {
     "space": true,
@@ -53,11 +55,6 @@
     ],
     "plugins": [
       "react"
-    ],
-    "settings": {
-      "react": {
-        "pragma": "h"
-      }
-    }
+    ]
   }
 }
